requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
firebase-admin>=6.4.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
google-cloud-firestore>=2.15.0